requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-ninja==1.1.0
python-decouple==3.8
qrcode[pil]==7.4.2
Pillow>=9.0.0
requests>=2.28.0
channels>=4.0.0
websockets>=11.0
django-debug-toolbar==4.2.0